pressure advance: do not smooth base extruder position, only advance
This eliminates extrusion errors (retraction issues, corner
distortions, etc.) caused by applying the smoothing kernel to the base
E position, which is not needed for or motivated by the pressure
advance model.

Fixes Danger Klipper issue 207 / upstream Klipper issue 4442.

Since the base position integrator is also used by changes to
kin_shaper for "input smoothing", that functionality is left in place,
but without the integration of the velocity term (the part needed for
PA) which is moved to its own function.